+++ GrowableGiantCrops/Framework/ModConfig.cs
@@ -36,6 +36,16 @@ public sealed class ModConfig
     }
 
     /// <summary>
+    /// Gets or sets a value indicating whether or not the shovel should do damage to monsters.
+    /// </summary>
+    public bool ShovelDoesDamage { get; set; } = true;
+
+    /// <summary>
+    /// Gets or sets a value indicating whether or not the shovel should be allowed the two hoe enchantments.
+    /// </summary>
+    public bool AllowHoeEnchantments { get; set; } = false;
+
+    /// <summary>
     /// Gets or sets a value indicating whether or not shops should have graphics.
     /// </summary>
     [GMCMSection("Shop", -10)]
@@ -56,6 +66,20 @@ public sealed class ModConfig
     [GMCMDefaultVector(6, 19)]
     [GMCMSection("Shop", -10)]
     public Vector2 ResourceShopLocation { get; set; } = new(6, 19);
+
+    private int maxGiantCropsSold = 5;
+
+    /// <summary>
+    /// Gets or sets the maximum number of giant crops that will be sold in the giant crop store,
+    /// if perfection has not been reached.
+    /// </summary>
+    [GMCMRange(1, 25)]
+    [GMCMSection("Shop", -10)]
+    public int MaxGiantCropsSold
+    {
+        get => this.maxGiantCropsSold;
+        set => this.maxGiantCropsSold = Math.Max(value, 1);
+    }
 
     /// <summary>
     /// Gets or sets a value indicating whether or not NPCs should trample placed resource clumps.