+++ SleepInWedding/ModEntry.cs
@@ -1,4 +1,5 @@
 ﻿using AtraCore.Framework.Caches;
+using AtraCore.Utilities;
 
 using AtraShared.ConstantsAndEnums;
 using AtraShared.Integrations;
@@ -89,7 +90,7 @@ internal sealed class ModEntry : Mod
             {
                 Game1.warpFarmer(new LocationRequest("Town", false, Game1.getLocationFromName("Town")), 5, 10, 0);
             }
-            else if (Game1.timeOfDay == Utility.ModifyTime(Config.WeddingTime, -30))
+            else if (Game1.timeOfDay == Utility.ModifyTime(Config.WeddingTime, -20))
             {
                 Game1.addHUDMessage(new HUDMessage(I18n.WeddingReminder(), HUDMessage.achievement_type));
             }
@@ -104,6 +105,13 @@ internal sealed class ModEntry : Mod
     /// <param name="e">Event args.</param>
     private void OnSaveLoad(object? sender, SaveLoadedEventArgs e)
     {
+        if (Context.IsSplitScreen && Context.ScreenId != 0)
+        {
+            return;
+        }
+
+        MultiplayerHelpers.AssertMultiplayerVersions(this.Helper.Multiplayer, this.ModManifest, this.Monitor, this.Helper.Translation);
+
         if (Context.IsMainPlayer && Config.TryRecoverWedding)
         {
             if (!Game1.canHaveWeddingOnDay(Game1.dayOfMonth, Game1.currentSeason))