+++ CritterRings/docs/changelog.md
@@ -1,5 +1,9 @@
 ﻿Changelog
 ==============
 
+## Version 0.1.1
+* Fix frogs spawning on beach when `SaltwaterFrogs` was disabled.
+* Added Russian translation, thanks to pavlo2906!
+
 ## Version 0.1.0
 * Initial Release.